refactor: Use interactive spec in advice for vscroll
Following recommendation in `called-interactively-p` avoid using `interactive`
parameter for such a call. Use an interactive spec of advice to preppend args
with `pk/interactive` symbol instead. This symbol is later detected, acted
upon, and stripped from args in the actual advice.

This remove the need to check for point visibility before calling checking
`call-interactively-p'.  The former is relatively fast compared to the latter.
This came up from investigating slowness in `helm-M-x'.  When affixation is
computed for the latter, it calls `substitute-command-keys' which in turn calls
forward-char'.